Se agrega condicion que no pueda eliminar usuarios con rol de director

+++ simulador-banca-main/Frontend/src/Components/DashBoard/Components/Director/CrearUsuario.jsx
@@ -15,11 +15,8 @@ export const CrearUsuario = () => {
   const [modalData, setModalData] = useState(null);
 
   const [showModal, setShowModal] = useState(false);
-  // Abrir Modal
   const [openModal1, setOpenModal] = useState(false);
-
   const [openConsing, setOpenConsing] = useState(false);
-  const [amount, setAmount] = useState("");
 
   const { user } = useAuth();
 
@@ -50,6 +47,7 @@ export const CrearUsuario = () => {
       if (response.ok) {
         const data = await response.json();
         setIdEmpleadoDetails(data);
+        return data;
       } else {
         console.error("Error fetching user info:", response.status);
       }
@@ -58,45 +56,31 @@ export const CrearUsuario = () => {
     }
   };
 
-  const AddUser = async (data) => {
-    try {
-      const response = await fetch("http://localhost:3000/add_user", {
-        method: "POST",
-        headers: {
-          "Content-type": "application/json",
-        },
-        body: JSON.stringify({
-          name: data.name,
-          password: data.password,
-          rol: data.rol,
-        }),
-      });
-      console.log(data);
-    } catch (error) {
-      console.error("Error al añadir usuario:", error);
-    }
-  };
-
   const eliminarUsuario = async (userId) => {
-    console.log(userId);
-    try {
-      const response = await fetch(
-        `http://localhost:3000/delete_user/${userId}`,
-        {
-          method: "DELETE",
+    const empleado = await fetchEmpleadoId(userId);
+
+    if (empleado.id_rol === 1) {
+      toast.error("Error: No puedes eliminar este usuario.");
+    } else {
+      try {
+        const response = await fetch(
+          `http://localhost:3000/delete_user/${userId}`,
+          {
+            method: "DELETE",
+          }
+        );
+        if (response.ok) {
+          toast.error("Usuario eliminado correctamente");
+          setTimeout(() => {
+            setForceUpdate((prev) => !prev);
+          }, 1000);
+        } else {
+          console.error("Error al eliminar usuario");
+          alert("Error al eliminar usuario");
         }
-      );
-      if (response.ok) {
-        toast.error("Usuario eliminado correctamente");
-        setTimeout(() => {
-          setForceUpdate((prev) => !prev);
-        }, 1000);
-      } else {
-        console.error("Error al eliminar usuario");
-        alert("Error al eliminar usuario");
+      } catch (error) {
+        console.error("Error en el servidor", error);
       }
-    } catch (error) {
-      console.error("Error en el servidor", error);
     }
   };
 
@@ -111,10 +95,6 @@ export const CrearUsuario = () => {
 
     // Formatea el costo usando la configuración especificada.
     return formatter.format(saldo);
-  };
-
-  const countUsers = () => {
-    return empleadoDetails.length;
   };
 
   const handleEmpleado = (idEmpleado) => {
@@ -151,7 +131,7 @@ export const CrearUsuario = () => {
                 Miembros del equipo
               </h2>
               <span className="px-2 py-1 text-xs text-blue-600 bg-blue-100 rounded-full dark:bg-gray-800 dark:text-blue-400">
-                {countUsers()} empleado(s)
+                {empleadoDetails.length} empleado(s)
               </span>
             </div>
             {user?.id_rol === 1 && (
@@ -264,88 +244,6 @@ export const CrearUsuario = () => {
                       <tbody className="bg-white divide-y divide-gray-200 dark:divide-gray-700 dark:bg-gray-900">
                         {empleadoDetails?.map((date) => (
                           <React.Fragment key={date.id_empleado}>
-                            {/* <tr>
-                              <td className="px-4 py-4 text-sm font-medium text-gray-700 dark:text-gray-200 whitespace-nowrap">
-                                <div className="inline-flex items-center gap-x-3">
-                                  <span># {date.id_empleado}</span>
-                                </div>
-                              </td>
-                              <td className="px-4 py-4 text-sm font-medium text-gray-700 whitespace-nowrap">
-                                <div className="inline-flex items-center gap-x-3">
-                                  <div className="flex items-center gap-x-2">
-                                    <img
-                                      className="object-cover w-10 h-10 rounded-full"
-                                      src="https://images.unsplash.com/photo-1535713875002-d1d0cf377fde?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=880&q=80"
-                                      alt=""
-                                    />
-                                    <div>
-                                      <h2 className="font-medium text-gray-800 dark:text-white ">
-                                        {date.username}
-                                      </h2>
-                                    </div>
-                                  </div>
-                                </div>
-                              </td>
-                              <td className="px-12 py-4 text-sm font-medium text-gray-700 whitespace-nowrap">
-                                <div className="inline-flex items-center px-3 py-1 rounded-full gap-x-2 bg-emerald-100/60 dark:bg-gray-800">
-                                  <span className="h-1.5 w-1.5 rounded-full bg-emerald-500"></span>
-
-                                  <h2 className="text-sm font-normal text-emerald-500">
-                                    Activo
-                                  </h2>
-                                </div>
-                              </td>
-                              <td className="px-4 py-4 text-sm text-gray-500 dark:text-gray-300 whitespace-nowrap">
-                                {date.id_rol === 1 && <>Director</>}
-                                {date.id_rol === 2 && <>Asesor</>}
-                                {date.id_rol === 3 && <>Cajero</>}
-                                {date.id_rol === 4 && <>Cajero Principal</>}
-                              </td>
-
-                              <td className="px-4 py-4 text-sm whitespace-nowrap">
-                                <div className="flex items-center gap-x-6">
-                                  <button
-                                    className="text-gray-500 transition-colors duration-200 dark:hover:text-red dark:text-gray-300 hover:text-red focus:outline-none"
-                                    onClick={() =>
-                                      eliminarUsuario(date.id_empleado)
-                                    }
-                                  >
-                                    <svg
-                                      xmlns="http://www.w3.org/2000/svg"
-                                      fill="none"
-                                      viewBox="0 0 24 24"
-                                      strokeWidth={1.5}
-                                      stroke="currentColor"
-                                      className="w-5 h-5"
-                                    >
-                                      <path
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                        d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0"
-                                      />
-                                    </svg>
-                                  </button>
-
-                                  <button className="text-gray-500 transition-colors duration-200 dark:hover:text-yellow-500 dark:text-gray-300 hover:text-yellow-500 focus:outline-none">
-                                    <svg
-                                      xmlns="http://www.w3.org/2000/svg"
-                                      fill="none"
-                                      viewBox="0 0 24 24"
-                                      strokeWidth={1.5}
-                                      stroke="currentColor"
-                                      className="w-5 h-5"
-                                    >
-                                      <path
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                        d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10"
-                                      />
-                                    </svg>
-                                  </button>
-                                </div>
-                              </td>
-                            </tr> */}
-
                             <tr>
                               <td className="px-4 py-4 text-sm font-medium text-gray-700 dark:text-gray-200 whitespace-nowrap">
                                 <div className="w-full inline-flex justify-center items-center gap-x-3">
